back-end/requirements.txt
flask
flask-jwt-extended
flask-socketio
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask_cors
